# website/sitio/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Enlyce.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/interacciones/lead/{leadId}": {
      "get": {
        "tags": [
          "Interacciones"
        ],
        "operationId": "ObtenerInteraccionesPorLead",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interaccion"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}": {
      "get": {
        "tags": [
          "Leads"
        ],
        "operationId": "GetLeadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLeadByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads": {
      "post": {
        "tags": [
          "Leads"
        ],
        "operationId": "CreateLead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLeadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/visitas/lead/{leadId}": {
      "get": {
        "tags": [
          "Visitas"
        ],
        "operationId": "ObtenerVisitasPorLead",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Visita"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLeadRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "fuente": {
            "type": "string",
            "nullable": true
          },
          "autorizacionDatos": {
            "type": "boolean"
          },
          "tipoOperacion": {
            "type": "string",
            "nullable": true
          },
          "ownerService": {
            "type": "string",
            "nullable": true
          },
          "publicationId": {
            "type": "string",
            "nullable": true
          },
          "canal": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateLeadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "ownerService": {
            "type": "string",
            "nullable": true
          },
          "publicationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "esContactoRepetido": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "GetLeadByIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "fuente": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "motivoCierre": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaUltimoContacto": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "autorizacionDatos": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "etapaPipeline": {
            "type": "string",
            "nullable": true
          },
          "tipoOperacion": {
            "type": "string",
            "nullable": true
          },
          "ownerService": {
            "type": "string",
            "nullable": true
          },
          "publicationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Interaccion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "leadId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "asesorId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "tipo": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "resumen": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "Visita": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "leadId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "inmuebleId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "asesorId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "fechaProgramada": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "fechaRealizada": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "feedback": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "estado": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}